Move Either extensions to separate file
